# optimize evm interpreter div operation