Below is the bench result of u256 div:

This benchmark compares two implementations of the `U256` division operation:

1. **Common**: directly uses the built-in `U256` division.
2. **Optimized**: uses bitwise operations for a set of specific divisors (ten in total); for all other cases, it falls back to the standard `U256` division.

### Benchmark results (on my local machine):

1. For general divisors, the **Optimized** implementation is actually **40% slower** than the **Common** implementation.
2. For specific divisors with early-match conditions (e.g., `2`, `256`), the **Optimized** version performs up to **4x faster** than the **Common** version.
   However, for late-match cases such as `TWO_POW_248`, the Optimized version is still **slower** than the Common one.

### Conclusion:

Given the low probability of encountering these specific divisors in real-world scenarios, the **standard division implementation is overall faster** and more suitable.

```sh
     Running benches/u256.rs (target/release/deps/u256-f9b156515e931e82)
U256 Division Special Cases/Common/2
                        time:   [17.544 ns 17.574 ns 17.610 ns]
                        change: [−0.6382% −0.0209% +0.5046%] (p = 0.94 > 0.05)
                        No change in performance detected.
Found 21 outliers among 100 measurements (21.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  19 (19.00%) high severe
U256 Division Special Cases/Optimized/2
                        time:   [4.0172 ns 4.1765 ns 4.4112 ns]
                        change: [−1.2104% +1.1830% +4.5175%] (p = 0.49 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe
U256 Division Special Cases/Common/256
                        time:   [17.547 ns 17.993 ns 18.947 ns]
                        change: [+0.0890% +1.1660% +3.0384%] (p = 0.17 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe
U256 Division Special Cases/Optimized/256
                        time:   [4.2777 ns 4.4238 ns 4.6603 ns]
                        change: [+0.0718% +2.4526% +5.4810%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 16 outliers among 100 measurements (16.00%)
  7 (7.00%) high mild
  9 (9.00%) high severe
U256 Division Special Cases/Common/72057594037927936
                        time:   [3.3135 ns 3.3613 ns 3.4529 ns]
                        change: [+0.3336% +1.8456% +4.1620%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 16 outliers among 100 measurements (16.00%)
  5 (5.00%) high mild
  11 (11.00%) high severe
U256 Division Special Cases/Optimized/72057594037927936
                        time:   [6.1175 ns 6.1241 ns 6.1319 ns]
                        change: [−8.7028% −4.8431% −1.7582%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  9 (9.00%) high mild
  7 (7.00%) high severe

U256 Division Normal Cases/Common/3
                        time:   [17.646 ns 18.133 ns 18.876 ns]
                        change: [+0.3264% +1.5986% +3.7856%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 16 outliers among 100 measurements (16.00%)
  4 (4.00%) high mild
  12 (12.00%) high severe
U256 Division Normal Cases/Optimized/3
                        time:   [24.920 ns 25.789 ns 26.944 ns]
                        change: [−0.0254% +1.8161% +4.0942%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  10 (10.00%) high severe
U256 Division Normal Cases/Common/6
                        time:   [17.579 ns 17.615 ns 17.654 ns]
                        change: [+0.0715% +0.4856% +0.9559%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
U256 Division Normal Cases/Optimized/6
                        time:   [24.819 ns 24.869 ns 24.932 ns]
                        change: [−0.5008% −0.2083% +0.0865%] (p = 0.18 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  10 (10.00%) high severe
```